Changed lighting

+++ src/block/mod.rs
@@ -15,7 +15,7 @@ pub struct Block {
     pub texture_atlas_lookups: [TextureAtlasIndex; 6],
     pub transparent: bool,
     pub light_intensity: u8,
-    pub color: [f32; 3]
+    pub color: [u8; 3],
 }
 
 /// Stores a direction on a block. Useful for things like furnaces that have a different front face, face culling and hoppers that move items down or sideways.
